flatten structure and use npm scripts
