--- GruntFile.js
@@ -1,2 +1,0 @@
-//github.com/ryanve/universal#grunt
-module.exports = require('./node_modules/universal/GruntFile');
