+++ test.js
@@ -1,11 +1,11 @@
-(function(root) {
+!function(root) {
   function contains(str, needle) {
     return !!~str.indexOf(needle);
   }
 
   var common = typeof module != 'undefined' && !!module.exports;
   var aok = common ? require('aok') : root.aok;
-  var eol = common ? require('../src') : root.eol;
+  var eol = common ? require('./') : root.eol;
   var platform = typeof process != 'undefined' && process.platform;
   var meths = ['lf', 'cr', 'crlf', 'auto'];
   var chars = ['\n', '\r', '\r\n', 'win32' === platform ? '\r\n' : '\n'];
@@ -29,4 +29,4 @@
   aok('auto agress', 2 === aok.pass(meths, function(method) {
     return eol.auto(sample) === eol[method](sample);
   }));
-}(this));
+}(this);
