--- test/index.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<title>Tests</title>
-<style>
-  html { font-family:sans-serif; background:#efefef; color:#111 }
-  body { width:96%; margin:auto; font-size:1em; line-height:1.6 }
-  pre, code, kbd { font-family:monospace; -moz-tab-size:2; -o-tab-size:2; tab-size:2 }
-  h3 b { font-weight:700 }
-  h3, pre b { font-weight:400 }
-  pre { padding:0 .5em; white-space:pre-line }
-  pre span { white-space:pre-wrap }
-  pre b { color:#008080 }
-  pre small { font-size:inherit; font-family:inherit; color:#707070 }
-  kbd { display:inline-block; padding:.2em } 
-  pre kbd { padding:0 }
-</style>
-
-<h1>To run tests</h1>
-<h3>Install <a href="../package.json">devDependencies</a></h3>
-<pre><span><b>$</b> <kbd>npm install</kbd></span></pre>
-<h3><b>Browser</b>: open the console</h3>
-<h3><b>CLI</b>: run tests via <a href="https://github.com/ryanve/aok#grunt-aok">grunt aok</a></h3>
-<pre><span><b>$</b> <kbd>grunt aok</kbd></span></pre>
-<h1>To edit tests</h1>
-<p>Edit <a href="https://github.com/ryanve/aok#readme">aok</a> tests in <a href="index.js">index.js</a></p>
-
-<script src="../src/index.js"></script>
-<script src="../node_modules/aok/aok.min.js"></script>
-<script src="./index.js"></script>
